Enforce #pragma once behavior
I believe that it's going to be a lot safer to enforce the behavior of
the C/C++ #pragma once directive by default that only allows a file to
be included once an no more. If I ever have a use case where I need to
include something multiple times I can work on dealing with it then, but
this should prevent errors and also seems to prevent recursive includes
from happening.

If you have a file `alpha.txt` with these contents:
    alpha
    #include "bravo.txt"

And a file `bravo.txt` with these contents:
    bravo
    #include "alpha.txt"

And you run the script on alpha.txt, you will end up with the following:
    alpha
    bravo

The script will stop because bravo's include of alpha will essentially
be a no-op, preventing a recursive loop.